# Add an initial projective approximation for linescan camera