--- a/environment.yml
+++ b/environment.yml
@@ -8,3 +8,4 @@ dependencies:
   - ale
   - csm
   - nlohmann_json
+  - eigen