--- a/include/usgscsm/UsgsAstroLsSensorModel.h
+++ b/include/usgscsm/UsgsAstroLsSensorModel.h
@@ -973,8 +973,16 @@ class UsgsAstroLsSensorModel : public csm::RasterGM,
   void computeLinearApproximation(const csm::EcefCoord& gp,
                                   csm::ImageCoord& ip) const;
 
-  // Initial setup of the linear approximation
-  void setLinearApproximation();
+  // Create the linear approximation to be used at each ground point
+  void createLinearApproximation();
+
+  // The projective approximation for the sensor model is used as the starting point
+  // for iterative rigorous calculations.
+  void computeProjectiveApproximation(const csm::EcefCoord& gp,
+                                      csm::ImageCoord& ip) const;
+  
+  // Create the projective approximation to be used at each ground point
+  void createProjectiveApproximation();
 
   // Compute the determinant of a 3x3 matrix
   double determinant3x3(double mat[9]) const;
@@ -991,7 +999,7 @@ class UsgsAstroLsSensorModel : public csm::RasterGM,
   std::vector<double>
       _no_adjustment;  // A vector of zeros indicating no internal adjustment
 
-  // The following support the linear approximation of the sensor model
+  // Store here the linear approximation of the sensor model
   double _u0;
   double _du_dx;
   double _du_dy;
@@ -1000,7 +1008,12 @@ class UsgsAstroLsSensorModel : public csm::RasterGM,
   double _dv_dx;
   double _dv_dy;
   double _dv_dz;
-  bool _linear;  // flag indicating if linear approximation is useful.
+
+  // Store here the projective approximation of the sensor model
+  std::vector<double> m_projTransCoeffs; 
+
+  // Flag indicating if an initial approximation of some kind is used
+  bool m_useApproxInitTrans;
 };
 
 #endif  // INCLUDE_USGSCSM_USGSASTROLSSENSORMODEL_H_