--- a/include/usgscsm/UsgsAstroLsSensorModel.h
+++ b/include/usgscsm/UsgsAstroLsSensorModel.h
@@ -968,14 +968,6 @@ class UsgsAstroLsSensorModel : public csm::RasterGM,
       const std::vector<double>& adj)      // Parameter Adjustments for partials
       const;
 
-  // The linear approximation for the sensor model is used as the starting point
-  // for iterative rigorous calculations.
-  void computeLinearApproximation(const csm::EcefCoord& gp,
-                                  csm::ImageCoord& ip) const;
-
-  // Create the linear approximation to be used at each ground point
-  void createLinearApproximation();
-
   // The projective approximation for the sensor model is used as the starting point
   // for iterative rigorous calculations.
   void computeProjectiveApproximation(const csm::EcefCoord& gp,
@@ -983,9 +975,6 @@ class UsgsAstroLsSensorModel : public csm::RasterGM,
   
   // Create the projective approximation to be used at each ground point
   void createProjectiveApproximation();
-
-  // Compute the determinant of a 3x3 matrix
-  double determinant3x3(double mat[9]) const;
   
   // A function whose value will be 0 when the line a given ground
   // point projects into is found. The obtained line will be
@@ -996,23 +985,12 @@ class UsgsAstroLsSensorModel : public csm::RasterGM,
   
   csm::NoCorrelationModel _no_corr_model;  // A way to report no correlation
                                            // between images is supported
-  std::vector<double>
-      _no_adjustment;  // A vector of zeros indicating no internal adjustment
-
-  // Store here the linear approximation of the sensor model
-  double _u0;
-  double _du_dx;
-  double _du_dy;
-  double _du_dz;
-  double _v0;
-  double _dv_dx;
-  double _dv_dy;
-  double _dv_dz;
+  std::vector<double> _no_adjustment;  // A vector of zeros indicating no internal adjustment
 
   // Store here the projective approximation of the sensor model
   std::vector<double> m_projTransCoeffs; 
 
-  // Flag indicating if an initial approximation of some kind is used
+  // Flag indicating if an initial approximation is used
   bool m_useApproxInitTrans;
 };