--- a/tests/CMakeLists.txt
+++ b/tests/CMakeLists.txt
@@ -9,7 +9,8 @@ add_executable(runCSMCameraModelTests
                DistortionTests.cpp
                SarTests.cpp
                ISDParsingTests.cpp
-               UtilitiesTests.cpp)
+               UtilitiesTests.cpp
+               EigenUtilitiesTests.cpp)
 if(WIN32)
   option(CMAKE_USE_WIN32_THREADS_INIT "using WIN32 threads" ON)
   option(gtest_disable_pthreads "Disable uses of pthreads in gtest." ON)