@oleg-alexandrov 2022-05-03
I added a unit test, removed the debug logic, wiped the old transform, and all the builds and tests pass.