@oleg-alexandrov 2022-05-03
I think linking to Eigen is not needed because a header file has the implementation. I used the same logic as what is in ALE.