@oleg-alexandrov 2022-05-03
The "target_link_libraries" logic should not help. Its doc says: "Specify libraries or flags to use when linking a given target and/or its dependents. ". This should not affect the problem I was seeing with finding headers which was happening before linking. 

What you will need, I think is "target_include_directories". Normally even this should not be needed, as "find_package" should already set the header, but that was not happening.

Anyhow, I spent a couple of hours yesterday and today to just get things to build. Now things work. If you are feeling adventurous and want to revisit this, one needs to ensure that the build works with both internal and external Eigen (since different logic is used for each), and then see if anything breaks after pushing things.